requirements.txt
feedparser>=6.0.11
Flask>=2.0.2
numpy>=2.0.0
scikit-learn>=1.5.1
sqlitedict>=2.1.0
sqlalchemy>=1.4.42
psycopg2>=2.9.5
python-dotenv>=1.0.1
